rename upsampler file, add option to just combine csv files for a channel in addition to upsampling them
